Continue input penjualan progress (Error on multiple delete)

+++ database/migrations/2018_02_01_073223_create_anggotas_table.php
@@ -19,7 +19,7 @@ class CreateAnggotasTable extends Migration
             $table->string('nama',30);
             $table->string('departemen',30);
             $table->string('posisi',30);
-            $table->bigInteger('totalSimpanan');
+            $table->integer('totalSimpanan');
             $table->timestamps();
         });
     }
