+++ database/migrations/2018_02_02_010643_create_barangs_table.php
@@ -16,7 +16,7 @@ class CreateBarangsTable extends Migration
         Schema::create('barangs', function (Blueprint $table) {
             $table->increments('id');
             $table->string('nama',30)->unique();
-            $table->bigInteger('harga');
+            $table->integer('harga');
             $table->timestamps();
         });
     }
